# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "DailyRequest" : {
        "properties" : {
          "day" : {
            "format" : "date",
            "type" : "string"
          },
          "dueTime" : {
            "format" : "time-local",
            "type" : "string"
          },
          "notes" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "position" : {
            "format" : "int32",
            "type" : "integer"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "TODO", "DOING", "DONE" ],
            "type" : "string"
          },
          "taskId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "DailyResponse" : {
        "properties" : {
          "day" : {
            "format" : "date",
            "type" : "string"
          },
          "dueTime" : {
            "format" : "time-local",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "notes" : {
            "type" : "string"
          },
          "position" : {
            "format" : "int32",
            "type" : "integer"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "TODO", "DOING", "DONE" ],
            "type" : "string"
          },
          "task" : {
            "$ref" : "#/components/schemas/TaskRef"
          },
          "title" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "TaskRef" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TaskResponse" : {
        "properties" : {
          "agentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "documentPath" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "phaseId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "OPEN", "IN_PROGRESS", "DONE" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TaskUpdateRequest" : {
        "properties" : {
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "priority" : {
            "minLength" : 1,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "priority", "title" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Every route needs `Authorization: Bearer <token>` (ADR-013). Every mutation of an existing resource needs `If-Match` with the ETag last read (ADR-009): absent is 428, stale is 412. Every error is `application/problem+json` with a stable `type` = `urn:ai-company-os:problem:<slug>` (ADR-007); branch on `type`, not on `title`.",
    "title" : "AI Company OS — control plane",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/daily" : {
      "get" : {
        "operationId" : "between",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DailyResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "daily-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DailyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DailyResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "daily-controller" ]
      }
    },
    "/api/daily/carry-over" : {
      "post" : {
        "operationId" : "carryOver",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DailyResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "daily-controller" ]
      }
    },
    "/api/daily/{id}" : {
      "delete" : {
        "operationId" : "remove",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "daily-controller" ]
      },
      "put" : {
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DailyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DailyResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "daily-controller" ]
      }
    },
...
    "/api/tasks/{id}" : {
      "get" : {
        "operationId" : "getTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      },
      "put" : {
        "operationId" : "updateTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ]
}